move desktop header menu into own file
